+++ packages/ai/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [0.45.4] - 2026-01-13
 
 ### Added